Move layout-view and layout-composeui into subpackages.
