+++ redwood-layout-composeui/src/main/kotlin/app/cash/redwood/layout/composeui/ComposeColumn.kt
@@ -13,25 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.layout
+package app.cash.redwood.layout.composeui
 
-import android.content.Context
-import android.view.View
+import androidx.compose.runtime.Composable
 import app.cash.redwood.LayoutModifier
 import app.cash.redwood.flexcontainer.FlexDirection
 import app.cash.redwood.layout.api.CrossAxisAlignment
 import app.cash.redwood.layout.api.MainAxisAlignment
 import app.cash.redwood.layout.api.Overflow
 import app.cash.redwood.layout.api.Padding
-import app.cash.redwood.layout.widget.Row
+import app.cash.redwood.layout.widget.Column
 import app.cash.redwood.widget.Widget
 
-internal class ViewRow(context: Context) : Row<View> {
-  private val container = ViewFlexContainer(context, FlexDirection.Row)
+internal class ComposeColumn : Column<@Composable () -> Unit> {
+  private val container = ComposeFlexContainer(FlexDirection.Column)
 
-  override val children: Widget.Children<View> get() = container.children
+  override val children: Widget.Children<@Composable () -> Unit> get() = container.children
 
-  override val value: View get() = container.view
+  override val value: @Composable () -> Unit get() = container.composable
 
   override var layoutModifiers: LayoutModifier
     get() = container.layoutModifiers
@@ -45,11 +44,11 @@ internal class ViewRow(context: Context) : Row<View> {
     container.overflow(overflow)
   }
 
-  override fun horizontalAlignment(horizontalAlignment: MainAxisAlignment) {
-    container.justifyContent(horizontalAlignment.toJustifyContent())
+  override fun horizontalAlignment(horizontalAlignment: CrossAxisAlignment) {
+    container.alignItems(horizontalAlignment.toAlignItems())
   }
 
-  override fun verticalAlignment(verticalAlignment: CrossAxisAlignment) {
-    container.alignItems(verticalAlignment.toAlignItems())
+  override fun verticalAlignment(verticalAlignment: MainAxisAlignment) {
+    container.justifyContent(verticalAlignment.toJustifyContent())
   }
 }
