+++ redwood-layout-composeui/src/test/kotlin/app/cash/redwood/layout/ComposeFlexContainerTest.kt
@@ -28,6 +28,7 @@ import app.cash.redwood.flexcontainer.AlignItems
 import app.cash.redwood.flexcontainer.FlexDirection
 import app.cash.redwood.flexcontainer.JustifyContent
 import app.cash.redwood.layout.api.Padding
+import app.cash.redwood.layout.composeui.ComposeFlexContainer
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
