+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/view/ViewColumn.kt
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.layout
+package app.cash.redwood.layout.view
 
-import androidx.compose.runtime.Composable
+import android.content.Context
+import android.view.View
 import app.cash.redwood.LayoutModifier
 import app.cash.redwood.flexcontainer.FlexDirection
 import app.cash.redwood.layout.api.CrossAxisAlignment
@@ -25,12 +26,12 @@ import app.cash.redwood.layout.api.Padding
 import app.cash.redwood.layout.widget.Column
 import app.cash.redwood.widget.Widget
 
-internal class ComposeColumn : Column<@Composable () -> Unit> {
-  private val container = ComposeFlexContainer(FlexDirection.Column)
+internal class ViewColumn(context: Context) : Column<View> {
+  private val container = ViewFlexContainer(context, FlexDirection.Column)
 
-  override val children: Widget.Children<@Composable () -> Unit> get() = container.children
+  override val children: Widget.Children<View> get() = container.children
 
-  override val value: @Composable () -> Unit get() = container.composable
+  override val value: View get() = container.view
 
   override var layoutModifiers: LayoutModifier
     get() = container.layoutModifiers
